Error checking on concat_dtbs.bbclass

+++ meta-emac/classes/concatdtbs.bbclass
@@ -7,6 +7,8 @@ python do_concat_dtbs() {
         kernel_file = f'{deploy_dir}/zImage'
         dtb_file = f'{deploy_dir}/{dtb}'
         concat_file = f'{deploy_dir}/zImage_{dtb.replace(".dtb", "")}'
+        if os.path.exists(concat_file):
+            bb.warn("Concatinated zImage_dtb file already existed and has be overwritten.")
         with open(kernel_file, "rb") as kf, open(dtb_file, "rb") as df, open(concat_file, "wb") as cf:
             cf.write(kf.read())
             cf.write(df.read())
@@ -19,7 +21,8 @@ python do_concat_dtbs_clean() {
     for dtb in dtbs:
         if dtb is '': continue
         concat_file = f'{deploy_dir}/zImage_{dtb.replace(".dtb", "")}'
-        os.remove(concat_file)
+        if os.path.exists(concat_file):
+            os.remove(concat_file)
 }
 
 addtask concat_dtbs_clean
